Data analysis for a loan service company. The company noticed a decrease in the number of repayments of loans and new users. The data i collected included details of users which amounted to over one hundred thousand users. The data contained details like loan status, borrower annual percentage rate, income range, monthly loan payment and others.
i worked as a data analyst for a loan service company. The company noticed a decrease in the number of repayments of loans and new users. The data i collected included details of users which amounted to over one hundred thousand users. The data contained details like loan status, borrower annual percentage rate, income range, monthly loan payment and others.
The Data was in a comma-separated values (CSV) file format. I found my answers by following a step-wise approach. Firstly, i performed wrangling on the dataset. I then selected a number of important parameters from the columns and added them to a data frame. Afterwards, i performed visualization and exploration on the new data frame from which i was able to generate relevant insights regarding the data.
I presented my findings in form of report and presentation which contained relevant visualization plots, explanations and insights derived.
A decision was made to make loans more easily accessible to users with high repayment rates and income range. A decision was also made to track the borrower annual percentage rates and monthly loan payment so as to provide a better experience to these users.